refactor: 네이버 지도 파일 위치 변경
